+++ rubrics/tool/pom.xml
@@ -240,102 +240,6 @@
                     <revisionOnScmFailure>Sakai</revisionOnScmFailure>
                 </configuration>
             </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-clean-plugin</artifactId>
-                <configuration>
-                    <filesets>
-                        <fileset>
-                            <directory>src/main/frontend/webcomponents/node_modules</directory>
-                            <followSymlinks>false</followSymlinks>
-                        </fileset>
-                        <fileset>
-                            <directory>src/main/frontend/webcomponents</directory>
-                            <includes>
-                                <include>**/*.js</include>
-                            </includes>
-                            <excludes>
-                                <exclude>js/**</exclude>
-                            </excludes>
-                            <followSymlinks>false</followSymlinks>
-                        </fileset>
-                    </filesets>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>com.github.eirslett</groupId>
-                <artifactId>frontend-maven-plugin</artifactId>
-                <version>1.8.0</version>
-                <configuration>
-                    <installDirectory>target</installDirectory>
-                    <workingDirectory>src/main/frontend/webcomponents</workingDirectory>
-                </configuration>
-                <executions>
-                    <execution>
-                        <id>install node and npm</id>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>install-node-and-npm</goal>
-                        </goals>
-                        <configuration>
-                            <nodeVersion>v12.16.1</nodeVersion>
-                            <npmVersion>6.14.2</npmVersion>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>npm install dependencies</id>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>npm</goal>
-                        </goals>
-                        <configuration>
-                            <arguments>
-                                ci --save-dev
-                            </arguments>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>npm empathise</id>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>npm</goal>
-                        </goals>
-                        <configuration>
-                            <arguments>
-                                run empathise -- --query=?version=${buildNumber}
-                            </arguments>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-				<artifactId>maven-resources-plugin</artifactId>
-				<version>2.6</version>
-				<executions>
-					<execution>
-						<id>copy-resources</id>
-						<phase>validate</phase>
-						<goals>
-							<goal>copy-resources</goal>
-						</goals>
-						<configuration>
-							<outputDirectory>${basedir}/target/classes/static/</outputDirectory>
-							<resources>          
-								<resource>
-									<directory>src/main/frontend</directory>
-                                    <filtering>true</filtering>
-                                    <excludes>
-                                        <exclude>webcomponents/js/**</exclude>
-                                        <exclude>webcomponents/node_modules/**</exclude>
-                                        <exclude>webcomponents/package.json</exclude>
-                                        <exclude>webcomponents/package-lock.json</exclude>
-                                    </excludes>
-								</resource>
-							</resources>              
-						</configuration>            
-					</execution>
-				</executions>
-            </plugin>
         </plugins>
     </build>
 </project>
