--- rubrics/tool/src/main/frontend/webcomponents/.gitignore
@@ -1,2 +1,0 @@
-*.js
-!js/**
