+++ rubrics/tool/src/main/resources/templates/index.html
@@ -4,7 +4,7 @@
   <head>
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8" />
     <script th:src="@{~/library/js/headscripts.js} + ${cdnQuery}"></script>
-    <script type="module" th:src="@{/webcomponents/sakai-rubrics-manager.js} + ${cdnQuery}"></script>
+    <script type="module" th:src="@{~/webcomponents/rubrics/sakai-rubrics-manager.js} + ${cdnQuery}"></script>
   </head>
   <body>
     <div class="portletBody">
