+++ samigo/samigo-app/src/webapp/jsf/author/item/shortAnswer.jsp
@@ -38,8 +38,8 @@
       <head><%= request.getAttribute("html.head") %>
       <title><h:outputText value="#{authorMessages.item_display_author}"/></title>
       <script src="/samigo-app/js/authoring.js"></script>
-      <script src="/rubrics-service/webcomponents/sakai-rubrics-utils.js<h:outputText value="#{itemauthor.CDNQuery}" />"></script>
-      <script type="module" src="/rubrics-service/webcomponents/rubric-association-requirements.js<h:outputText value="#{itemauthor.CDNQuery}" />"></script>
+      <script src="/webcomponents/rubrics/sakai-rubrics-utils.js<h:outputText value="#{itemauthor.CDNQuery}" />"></script>
+      <script type="module" src="/webcomponents/rubrics/rubric-association-requirements.js<h:outputText value="#{itemauthor.CDNQuery}" />"></script>
       </head>
 <body onload="countNum();<%= request.getAttribute("html.body.onload") %>">
