+++ samigo/samigo-app/src/webapp/jsf/delivery/item/deliverFileUpload.jsp
@@ -24,7 +24,7 @@
 include file for delivering file upload questions
 should be included in file importing DeliveryMessages
 --%>
-<script src="/rubrics-service/webcomponents/sakai-rubrics-utils.js<h:outputText value="#{studentScores.CDNQuery}" />"></script>
+<script src="/webcomponents/rubrics/sakai-rubrics-utils.js<h:outputText value="#{studentScores.CDNQuery}" />"></script>
 
 <h:outputText value="#{question.text} <br/>"  escape="false"/>
 <!-- ATTACHMENTS -->
