+++ webcomponents/tool/src/main/frontend/js/rubrics/rubric-association-requirements.js
