+++ webcomponents/tool/src/main/frontend/js/rubrics/rubrics-element.js
