+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-item-delete.js
