+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-creator-name.js
