+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-criteria-readonly.js
