+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-criterion-rating-edit.js
