+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-criterion-student.js
