+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-grading-button.js
