+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-modified-date.js
