+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-preview.js
