+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-site-title.js
