+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-student-preview-button.js
