+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubric-student.js
