+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubrics-helpers.js
