+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubrics-list.js
