+++ webcomponents/tool/src/main/frontend/js/rubrics/sakai-rubrics-shared-list.js
