+++ webcomponents/tool/src/main/frontend/js/rubrics/sharing-change-event.js
